requirements.txt

requests>=2.31.0
python-dotenv>=1.0.0

pandas>=2.1.0
numpy>=1.24.0
yfinance>=0.2.36


matplotlib>=3.8.0
mplfinance>=0.12.10b0


googletrans==3.1.0a0


tqdm>=4.66.0
